+++ packages/imperative/src/imperative/src/LoggingConfigurer.ts
@@ -9,13 +9,12 @@
 *
 */
 
+import * as path from "path";
 import { IConfigLogging } from "../../logger/src/doc/IConfigLogging";
 import { Logger } from "../../logger/src/Logger";
 import { LoggerConfigBuilder } from "../../logger/src/LoggerConfigBuilder";
 import { IImperativeConfig } from "./doc/IImperativeConfig";
 import { Console } from "../../console";
-import { isNullOrUndefined } from "util";
-import { IO } from "../../io/src/IO";
 import { IImperativeLoggingConfig } from "./doc/IImperativeLoggingConfig";
 import { ImperativeError } from "../../error/src/ImperativeError";
 import { ImperativeExpect } from "../../expect/src/ImperativeExpect";
@@ -89,9 +88,9 @@ export class LoggingConfigurer {
         /**
          * All remaining logs are created here
          */
-        if (!isNullOrUndefined(imperativeConfig.logging.additionalLogging)) {
+        if (imperativeConfig.logging.additionalLogging != null) {
             imperativeConfig.logging.additionalLogging.forEach((logConfig) => {
-                if (isNullOrUndefined(logConfig.apiName)) {
+                if (logConfig.apiName == null) {
                     throw new ImperativeError({
                         msg: "apiName is required for additionalLoggers",
                     });
@@ -160,11 +159,9 @@ export class LoggingConfigurer {
      */
     private static configureLoggerByKey(
         home: string, imperativeConfig: IImperativeConfig, loggingConfig: IConfigLogging, entryKey: string, configKey: string) {
-        if (!isNullOrUndefined(imperativeConfig.logging)) {
-            if (!isNullOrUndefined(imperativeConfig.logging[configKey])) {
-                loggingConfig = LoggingConfigurer.configureLoggerByKeyHelper(
-                    home, imperativeConfig.logging[configKey], loggingConfig, entryKey, configKey);
-            }
+        if (imperativeConfig.logging != null && imperativeConfig.logging[configKey] != null) {
+            loggingConfig = LoggingConfigurer.configureLoggerByKeyHelper(
+                home, imperativeConfig.logging[configKey], loggingConfig, entryKey, configKey);
         }
 
         return loggingConfig;
@@ -184,12 +181,12 @@ export class LoggingConfigurer {
      */
     private static configureLoggerByKeyHelper(home: string, impLogConfig: IImperativeLoggingConfig,
         loggingConfig: IConfigLogging, entryKey: string, configKey: string) {
-        if (!isNullOrUndefined(impLogConfig.logFile)) {
+        if (impLogConfig.logFile != null) {
             const fullLogFilePath = home +
                 LoggingConfigurer.normalizeDir(impLogConfig.logFile);
             loggingConfig.log4jsConfig.appenders[entryKey].filename = fullLogFilePath as any;
         }
-        if (!isNullOrUndefined(impLogConfig.level)) {
+        if (impLogConfig.level != null) {
             Console.validateLevel(impLogConfig.level);
             loggingConfig.log4jsConfig.categories[entryKey].level = impLogConfig.level;
         }
@@ -269,11 +266,11 @@ export class LoggingConfigurer {
      */
     private static buildLoggingDefaultsByKey(
         imperativeConfig: IImperativeConfig, key: string, apiName: string, category = apiName): IImperativeConfig {
-        if (isNullOrUndefined(imperativeConfig.logging)) {
+        if (imperativeConfig.logging == null) {
             imperativeConfig.logging = {};
             imperativeConfig.logging[key] = {apiName, category};
         } else {
-            if (isNullOrUndefined(imperativeConfig.logging[key])) {
+            if (imperativeConfig.logging[key] == null) {
                 imperativeConfig.logging[key] = {apiName, category};
             } else {
                 imperativeConfig.logging[key].apiName = apiName;
@@ -296,7 +293,7 @@ export class LoggingConfigurer {
         if (file[0] === "/" || file[0] === "\\") {
             return file;
         } else {
-            return IO.FILE_DELIM + file;
+            return path.posix.sep + file;
         }
     }
 }